give up previous design
