+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:theme="@style/Theme.AppCompat" >
         <activity
-            android:name=".QuizDroidModel"
+            android:name=".AppFrontPage"
             android:label="eeeeeed" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,15 +21,17 @@
             android:label="@string/topic_front_page" >
         </activity>
         <activity
-            android:name=".AppFrontPage"
+            android:name=".QuizDroidModel"
             android:label="@string/app_name" >
         </activity>
-
         <activity
             android:name=".QuizQuestions"
             android:label="@string/title_activity_quiz_questions" >
         </activity>
-
+        <activity
+            android:name=".Answer"
+            android:label="@string/title_activity_answer" >
+        </activity>
     </application>
 
 </manifest>
