+++ app/src/main/java/edu/washington/ruokua/quizdroid/AppFrontPage.java
@@ -3,8 +3,6 @@ package edu.washington.ruokua.quizdroid;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
@@ -13,13 +11,21 @@ import android.widget.ListView;
 
 public class AppFrontPage extends AppCompatActivity {
     // A list of different topics on which allow user to take quiz
-    private String[] topics;
+
+
+    private final int NUM_Math_QUESTIONS = 3;
+
+    private final int NUM_Physics_QUESTIONS = 0;
+
+    private final int NUM_MARVEL_QUESTIONS = 0;
+
+    private final String[] TOPICS = {"Math", "Physics", "Marvel Super Heroes"};
 
 
     //The view of topic
     private ListView topicList;
 
-
+    private String topic;
     /**
      * Display a List of topics on which allow user to take quiz
      *
@@ -33,22 +39,24 @@ public class AppFrontPage extends AppCompatActivity {
 
 
         Intent launchingIntent = getIntent();
-        topics = launchingIntent.getStringArrayExtra("topics");
 
         topicList = (ListView) findViewById(R.id.lstTopic);
 
         ArrayAdapter<String> items = new ArrayAdapter<String>(this,
-                R.layout.topic_list_item, topics);
+                R.layout.topic_list_item, TOPICS);
         topicList.setAdapter(items);
 
         topicList.setOnItemClickListener(new ListView.OnItemClickListener() {
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                Intent topicDesc = new Intent(AppFrontPage.this, TopicFrontPage.class);
+                topic = TOPICS[position];
 
-                Intent topicChoose = new Intent(AppFrontPage.this,
-                        QuizDroidModel.class);
-                topicChoose.putExtra("topicIndex", position);
+                topicDesc.putExtra("topic", topic);
 
-                startActivity(topicChoose);
+                int numQuestion = getNumProblem();
+                topicDesc.putExtra("numProblem", numQuestion);
+
+                startActivity(topicDesc);
 
 
             }
@@ -58,25 +66,19 @@ public class AppFrontPage extends AppCompatActivity {
     }
 
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main, menu);
-        return true;
-    }
 
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
 
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_settings) {
-            return true;
+    private int getNumProblem() {
+        switch (topic) {
+            case "Math":
+                return NUM_Math_QUESTIONS;
+            case "Physics":
+                return NUM_Physics_QUESTIONS;
+            default:
+
+                return NUM_MARVEL_QUESTIONS;
         }
-
-        return super.onOptionsItemSelected(item);
     }
+
+
 }
