+++ app/src/main/java/edu/washington/ruokua/quizdroid/QuestionList.java
@@ -1,0 +1,78 @@
+package edu.washington.ruokua.quizdroid;
+
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Created by ruokua on 4/27/15.
+ */
+public class QuestionList implements Serializable {
+    private List<String> desc;
+    private List<List<String>> option;
+    private List<Integer> answer;
+    private int questionContain;
+    private int questionNum;
+    private int selectAnswer;
+
+    public int getSelectAnswer() {
+        return selectAnswer;
+    }
+
+    public void setSelectAnswer(int selectAnswer) {
+        this.selectAnswer = selectAnswer;
+    }
+
+    public QuestionList(List<String> desc, List<List<String>> option, List<Integer> answer) {
+        if (desc.size() != option.size() || desc.size() != answer.size()) {
+            throw new IllegalArgumentException();
+        }
+        this.desc = desc;
+        this.option = option;
+        this.answer = answer;
+        this.questionContain = desc.size();
+    }
+
+
+    public int getQuestionNum() {
+        return questionNum;
+    }
+
+        
+    public int getQuestionContain() {
+        return questionContain;
+    }
+
+
+    public void nextQuestion() {
+        questionNum++;
+    }
+
+    public boolean hasNextQuestion() {
+        return questionNum < questionContain;
+    }
+
+
+    public String getDesc() {
+        if (questionNum < 0 || questionNum > questionContain) {
+            return "";
+        }
+        return desc.get(questionNum);
+    }
+
+    public List<String> getOption() {
+        if (questionNum < 0 || questionNum > questionContain) {
+            return new ArrayList<String>();
+        }
+        return option.get(questionNum);
+    }
+
+    public int getAnswer() {
+        if (questionNum < 0 || questionNum > questionContain) {
+            return 0;
+        }
+        return answer.get(questionNum);
+    }
+}
+
+
