+++ app/src/main/res/layout/activity_answer.xml
@@ -1,0 +1,81 @@
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    tools:context="edu.washington.austindg.quizdroid.AnswerSummaryActivity"
+    android:gravity="center"
+    android:background="#D2D93B">
+    <LinearLayout
+        android:orientation="vertical"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true"
+        android:weightSum="1"
+        android:id="@+id/linearLayout">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Correct/Incorrect"
+            android:id="@+id/correct_incorrect_text"
+            android:gravity="center"
+            android:textSize="24dp"
+            android:paddingBottom="15dp"
+            android:textColor="#ff212121" />
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Given Answer:"
+            android:textSize="16dp"
+            />
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/answerGiven"
+            android:layout_weight="0.5"
+            android:paddingLeft="20dp" />
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Correct Answer:"
+            android:id="@+id/textView3"
+            android:textSize="16dp"
+            android:paddingTop="10dp" />
+        <TextView
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/answer"
+            android:layout_weight="0.5"
+            android:textSize="14dp"
+            android:paddingLeft="20dp" />
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="vertical"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_centerVertical="true">
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="vertical"
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true">
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/question_ratio_text"
+            android:textSize="16dp"
+            android:gravity="center"
+           />
+        <Button
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:text="Next"
+            android:id="@+id/nextButton" />
+    </LinearLayout>
+</RelativeLayout>
