+++ app/src/main/res/layout/activity_quiz_questions.xml
@@ -31,7 +31,7 @@
             android:id="@+id/option0"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:tag="0" />
+            android:tag= "0" />
 
         <RadioButton
             android:id="@+id/option1"
