+++ app/src/main/res/layout/topic_overview.xml
@@ -28,9 +28,8 @@
         android:paddingBottom="@dimen/activity_vertical_margin"
         android:paddingLeft="@dimen/activity_horizontal_margin"
         android:paddingRight="@dimen/activity_horizontal_margin"
-        android:text="the branch of science concerned with the nature and properties of matter and energy. The subject matter of physics, distinguished from that of chemistry and biology, includes mechanics, heat, light and other radiation, sound, electricity, magnetism, and the structure of atoms.
 
-"
+
         android:textSize="15dp"
         android:textColor="#068481"
 
@@ -69,7 +68,7 @@
         android:id="@+id/btnStart"
         android:text="@string/quiz_begin"
 
-        style="@style/btnStyleSeagull"
+
         android:layout_marginTop="45dp"
         android:layout_below="@+id/textView2"
         android:layout_alignParentLeft="true"
