+++ app/src/main/res/values/strings.xml
@@ -11,6 +11,7 @@
 
     <string name="title_activity_quiz_questions">QuizQuestions</string>
     <string name="title_activity_quiz_droid_model">QuizDroidModel</string>
+    <string name="title_activity_answer">Answer</string>
 
 
 </resources>
